test(guest-download): split binary logging fixtures
